# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: الدمشقي ERP + e-commerce + POS + manufacturing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: "#/components/schemas/UserRole" }
                permissions:
                  type: array
                  items: { type: string }
                isActive: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      operationId: createCategory
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [catalog]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
    delete:
      operationId: deleteCategory
      tags: [catalog]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity]
              properties:
                productId: { type: string }
                quantity: { type: integer, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
    delete:
      operationId: removeCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
  /checkout:
    post:
      operationId: checkout
      tags: [checkout]
      summary: Convert cart into an online sales order (cash on delivery)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckoutInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }
...
  /content-blocks:
    get:
      operationId: listContentBlocks
      tags: [cms]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ContentBlock" }
  /content-blocks/{key}:
    get:
      operationId: getContentBlock
      tags: [cms]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContentBlock" }
    put:
      operationId: upsertContentBlock
      tags: [cms]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentBlockInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContentBlock" }
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [admin, manager, production, store, cashier, customer]
...
    AdminUser:
      type: object
      required: [id, email, role, isActive, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        imageUrl: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        permissions:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    Category:
      type: object
      required: [id, slug, nameAr, sortOrder, createdAt]
      properties:
        id: { type: string }
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        imageUrl: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
    CategoryInput:
      type: object
      required: [slug, nameAr]
      properties:
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        imageUrl: { type: string }
        sortOrder: { type: integer }
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, preparing, ready, out_for_delivery, completed, cancelled]
    SalesOrder:
      type: object
      required: [id, orderNumber, channel, status, subtotalMinor, totalMinor, currency, items, createdAt]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        channel: { type: string, enum: [pos, online] }
        status: { $ref: "#/components/schemas/OrderStatus" }
        customerId: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        customerEmail: { type: string }
        deliveryAddress: { type: string }
        deliveryNotes: { type: string }
        paymentMethod: { type: string, enum: [cash, card, cod, online] }
        subtotalMinor: { type: integer }
        discountMinor: { type: integer }
        deliveryFeeMinor: { type: integer }
        totalMinor: { type: integer }
        currency: { type: string }
        cashierId: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/SalesOrderItem" }
        createdAt: { type: string, format: date-time }
    SalesOrderItem:
      type: object
      required: [id, productId, productNameAr, quantity, unitPriceMinor, lineTotalMinor]
      properties:
        id: { type: string }
        productId: { type: string }
        productNameAr: { type: string }
        quantity: { type: integer }
        unitPriceMinor: { type: integer }
        lineTotalMinor: { type: integer }
...
    CheckoutInput:
      type: object
      required: [customerName, customerPhone, deliveryAddress]
      properties:
        customerName: { type: string }
        customerPhone: { type: string }
        customerEmail: { type: string }
        deliveryAddress: { type: string }
        deliveryNotes: { type: string }
        paymentMethod: { type: string, enum: [cod, card, cash, bank_transfer, stripe, paypal] }
    Cart:
      type: object
      required: [items, subtotalMinor, currency]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        subtotalMinor: { type: integer }
        currency: { type: string }
    CartItem:
      type: object
      required: [id, productId, productNameAr, quantity, unitPriceMinor, lineTotalMinor]
      properties:
        id: { type: string }
        productId: { type: string }
        productSlug: { type: string }
        productNameAr: { type: string }
        productImageUrl: { type: string }
        quantity: { type: integer }
        unitPriceMinor: { type: integer }
        lineTotalMinor: { type: integer }
        unit: { type: string }
...
    ContentBlock:
      type: object
      required: [id, key, page, contentAr, updatedAt]
      properties:
        id: { type: string }
        key: { type: string }
        page: { type: string }
        titleAr: { type: string }
        contentAr: { type: string }
        imageUrl: { type: string }
        ctaLabel: { type: string }
        ctaHref: { type: string }
        updatedAt: { type: string, format: date-time }
    ContentBlockInput:
      type: object
      required: [page, contentAr]
      properties:
        page: { type: string }
        titleAr: { type: string }
        contentAr: { type: string }
        imageUrl: { type: string }
        ctaLabel: { type: string }
        ctaHref: { type: string }
...